Restructured backend code
