+++ backend/src/controllers/namespacesController.js
@@ -1,0 +1,18 @@
+import { k8sCoreApi } from "../config/kubernetesClient.js";
+
+// Get all namespaces
+export const getNamespaces = async (req, res) => {
+    try {
+        const response = await k8sCoreApi.listNamespace();
+
+        res.json({
+            items: response.items,
+        });
+    } catch (error) {
+        console.error("Error fetching namespaces:", error);
+        res.status(500).json({
+            error: "Failed to fetch namespaces",
+            details: error.message,
+        });
+    }
+};
