+++ backend/src/routes/namespacesRoutes.js
@@ -1,0 +1,9 @@
+import express from "express";
+import * as namespacesController from "../controllers/namespacesController.js";
+
+const router = express.Router();
+
+// Namespaces route
+router.get("/namespaces", namespacesController.getNamespaces);
+
+export default router;
